land and loan value mask fixed

+++ database/migrations/2021_12_06_160048_create_lands_table.php
@@ -18,8 +18,7 @@ class CreateLandsTable extends Migration
             $table->string('landAddress')->nullable();
             $table->string('landMap')->nullable();
             $table->string('landDetails')->nullable();
-            $table->float('landValue',10,2)->nullable();
-
+            $table->decimal('landValue',10,2)->nullable();
             $table->integer('ownerID')->nullable();
 
             $table->timestamps();
